+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Fix race condition between stack management and modal life
     - Bump pwix:i18n version requirement
     - Bump pwix:jquery-ui version requirement
+    - Obsoletize Modal.target() replaced with Modal.set() (todo #31)
 
 ### 1.6.0